Refactor OGC-INSPIRE services
