+++ arlas-server/src/main/java/io/arlas/server/app/ArlasServer.java
@@ -28,14 +28,14 @@ import com.smoketurner.dropwizard.zipkin.ZipkinFactory;
 import io.arlas.server.exceptions.*;
 import io.arlas.server.health.ElasticsearchHealthCheck;
 import io.arlas.server.ogc.csw.CSWHandler;
-import io.arlas.server.ogc.csw.ElasticCSWService;
+import io.arlas.server.ogc.csw.CSWService;
 import io.arlas.server.ogc.csw.writer.getrecords.AtomGetRecordsMessageBodyWriter;
 import io.arlas.server.ogc.csw.writer.getrecords.XmlGetRecordsMessageBodyWriter;
 import io.arlas.server.ogc.csw.writer.record.AtomRecordMessageBodyWriter;
 import io.arlas.server.ogc.csw.writer.record.XmlRecordMessageBodyBuilder;
 import io.arlas.server.ogc.csw.writer.record.XmlMDMetadataMessageBodyWriter;
-import io.arlas.server.ogc.wfs.WFSHandler;
 import io.arlas.server.ogc.wfs.WFSService;
+import io.arlas.server.ogc.wfs.WFSHandler;
 import io.arlas.server.rest.collections.ElasticCollectionService;
 import io.arlas.server.rest.explore.aggregate.AggregateRESTService;
 import io.arlas.server.rest.explore.aggregate.GeoAggregateRESTService;
@@ -180,7 +180,7 @@ public class ArlasServer extends Application<ArlasServerConfiguration> {
         if(configuration.arlasServiceWFSEnabled){
             LOGGER.info("WFS Service enabled");
             WFSHandler wfsHandler = new WFSHandler(configuration.wfsConfiguration, configuration.ogcConfiguration, configuration.inspireConfiguration);
-            environment.jersey().register(new WFSService(exploration, wfsHandler));
+            environment.jersey().register(new WFSService(exploration, configuration, wfsHandler));
         } else {
             LOGGER.info("WFS Service disabled");
         }
@@ -196,7 +196,7 @@ public class ArlasServer extends Application<ArlasServerConfiguration> {
         if (configuration.arlasServiceCSWEnabled) {
             LOGGER.info("CSW Service enabled");
             CSWHandler cswHandler = new CSWHandler(configuration.ogcConfiguration,configuration.cswConfiguration, configuration.inspireConfiguration);
-            environment.jersey().register(new ElasticCSWService(cswHandler, exploration,configuration));
+            environment.jersey().register(new CSWService(client, cswHandler,configuration));
         } else {
             LOGGER.info("CSW Service disabled");
         }
