+++ ogc-csw/src/main/java/io/arlas/server/ogc/csw/operation/getrecords/GetRecordsHandler.java
@@ -69,8 +69,6 @@ public class GetRecordsHandler {
                         JAXBElement<BriefRecordType> briefRecordType = cswHandler.cswFactory.createBriefRecord(briefRecord);
                         searchResultType.getAbstractRecord().add(briefRecordType);
                     }
-
-
                 });
                 break;
             case summary:
